requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
zipp==3.23.0
pymongo==4.7.1
bson==0.5.10
pandas==2.2.2
openpyxl==3.1.2
pytz==2024.1
python-dotenv==1.0.1
